data_ingestion/requirements.txt
time
scapy.all
ipcalc
json
datetime
google.cloud
logging